Move the polyfill to its own file.
