+++ src/libraries/System.Reflection.Metadata/src/System.Reflection.Metadata.csproj
@@ -27,8 +27,8 @@ The System.Reflection.Metadata library is built-in as part of the shared framewo
     <Compile Include="$(CommonPath)Interop\Windows\Kernel32\Interop.ReadFile_SafeHandle_IntPtr.cs"
              Link="Common\Interop\Windows\kernel32\Interop.ReadFile_SafeHandle_IntPtr.cs"
              Condition="'$(TargetFrameworkIdentifier)' != '.NETCoreApp'" />
-    <Compile Include="$(CommonPath)Polyfills\StreamPolyfills.cs"
-             Link="Common\Polyfills\StreamPolyfills.cs"
+    <Compile Include="$(CommonPath)Polyfills\StreamSpanPolyfills.cs"
+             Link="Common\Polyfills\StreamSpanPolyfills.cs"
              Condition="'$(TargetFrameworkIdentifier)' != '.NETCoreApp'" />
     <Compile Include="System\Reflection\Internal\Utilities\PinnedObject.cs" />
     <Compile Include="System\Reflection\Internal\Utilities\CriticalDisposableObject.cs" />
